improve locators and separate files creation in other method
Now the excel spreadsheet creation and image dowload are in a
new method called `generate_news_files`.

To fix StaleElementReferenceException errors when getting search results
a new method was created to get this results and a exception hadling was
added, if the StaleElementReferenceException triggers a new attempt is made
until the MAX_ATTEMPT (default 5) be reached.